fix(ai-daily): can not show entry paper click the daily entry link
- Introduced EntryModalPreview for displaying entry content in a modal.
- Added EntryMoreActions for handling additional actions related to entries.
- Implemented EntryToastPreview to show a toast notification for entries.
- Updated usePeekModal hook to integrate new components and streamline entry detail fetching.
